requirements.txt
tensorflow==2.13.0
streamlit
numpy
Pillow

